JupyterPlugin: Created separate files for JupyterHTMLView/JupyterTab
